relax the tree paths
