+++ Cargo.toml
@@ -19,6 +19,7 @@ ark-ff = { version = "^0.2.0", default-features = false }
 ark-ec = { version = "^0.2.0", default-features = false }
 ark-std = { version = "^0.2.0", default-features = false }
 ark-relations = { version = "^0.2.0", default-features = false }
+ark-serialize = { version = "^0.2.0", default-features = false, features = [ "derive" ] }
 
 blake2 = { version = "0.9", default-features = false }
 digest = "0.9"
