+++ src/encryption/mod.rs
@@ -36,94 +36,3 @@ pub trait AsymmetricEncryptionScheme {
         ciphertext: &Self::Ciphertext,
     ) -> Result<Self::Plaintext, Error>;
 }
-
-#[cfg(test)]
-mod test {
-    use crate::encryption::constraints::AsymmetricEncryptionGadget;
-    use ark_std::{test_rng, UniformRand};
-
-    use ark_ed_on_bls12_381::{constraints::EdwardsVar, EdwardsProjective as JubJub, Fq};
-
-    use crate::encryption::elgamal::{constraints::ElGamalEncGadget, ElGamal, Randomness};
-    use crate::encryption::AsymmetricEncryptionScheme;
-    use ark_r1cs_std::prelude::*;
-    use ark_relations::r1cs::ConstraintSystem;
-
-    #[test]
-    fn test_elgamal_encryption() {
-        let rng = &mut test_rng();
-
-        // setup and key generation
-        let parameters = ElGamal::<JubJub>::setup(rng).unwrap();
-        let (pk, sk) = ElGamal::<JubJub>::keygen(&parameters, rng).unwrap();
-
-        // get a random msg and encryption randomness
-        let msg = JubJub::rand(rng).into();
-        let r = Randomness::rand(rng);
-
-        // encrypt and decrypt the message
-        let cipher = ElGamal::<JubJub>::encrypt(&parameters, &pk, &msg, &r).unwrap();
-        let check_msg = ElGamal::<JubJub>::decrypt(&parameters, &sk, &cipher).unwrap();
-
-        assert_eq!(msg, check_msg);
-    }
-
-    #[test]
-    fn test_elgamal_gadget() {
-        let rng = &mut test_rng();
-
-        type MyEnc = ElGamal<JubJub>;
-        type MyGadget = ElGamalEncGadget<JubJub, EdwardsVar>;
-
-        // compute primitive result
-        let parameters = MyEnc::setup(rng).unwrap();
-        let (pk, _) = MyEnc::keygen(&parameters, rng).unwrap();
-        let msg = JubJub::rand(rng).into();
-        let randomness = Randomness::rand(rng);
-        let primitive_result = MyEnc::encrypt(&parameters, &pk, &msg, &randomness).unwrap();
-
-        // construct constraint system
-        let cs = ConstraintSystem::<Fq>::new_ref();
-        let randomness_var =
-            <MyGadget as AsymmetricEncryptionGadget<MyEnc, Fq>>::RandomnessVar::new_witness(
-                ark_relations::ns!(cs, "gadget_randomness"),
-                || Ok(&randomness),
-            )
-            .unwrap();
-        let parameters_var =
-            <MyGadget as AsymmetricEncryptionGadget<MyEnc, Fq>>::ParametersVar::new_constant(
-                ark_relations::ns!(cs, "gadget_parameters"),
-                &parameters,
-            )
-            .unwrap();
-        let msg_var =
-            <MyGadget as AsymmetricEncryptionGadget<MyEnc, Fq>>::PlaintextVar::new_witness(
-                ark_relations::ns!(cs, "gadget_message"),
-                || Ok(&msg),
-            )
-            .unwrap();
-        let pk_var =
-            <MyGadget as AsymmetricEncryptionGadget<MyEnc, Fq>>::PublicKeyVar::new_witness(
-                ark_relations::ns!(cs, "gadget_public_key"),
-                || Ok(&pk),
-            )
-            .unwrap();
-
-        // use gadget
-        let result_var =
-            MyGadget::encrypt(&parameters_var, &msg_var, &randomness_var, &pk_var).unwrap();
-
-        // check that result equals expected ciphertext in the constraint system
-        let expected_var =
-            <MyGadget as AsymmetricEncryptionGadget<MyEnc, Fq>>::OutputVar::new_input(
-                ark_relations::ns!(cs, "gadget_expected"),
-                || Ok(&primitive_result),
-            )
-            .unwrap();
-        expected_var.enforce_equal(&result_var).unwrap();
-
-        assert_eq!(primitive_result.0, result_var.c1.value().unwrap());
-        assert_eq!(primitive_result.1, result_var.c2.value().unwrap());
-        assert!(cs.is_satisfied().unwrap());
-    }
-}
